+++ App.js
@@ -1,18 +1,17 @@
 
 import React from 'react';
 import { View } from "react-native";
-import LoginScreen4 from './components/LoginScreen4';
+import LoginScreen7 from './components/LoginScreen7';
 import LoginScreen8 from './components/LoginScreen8';
 import LoginScreen10 from './components/LoginScreen10';
-import LoginScreen from './components/LoginScreen';
-// import LoginScreen1 from './components/login-pages/LoginScreen1';
-import LoginScreen1 from './components/LoginScreen1';
+import LoginScreen3 from './components/LoginScreen3';
+import LoginScreen4 from './components/LoginScreen4';
 
 
 export default App = () => {
     return (
         <View style={{flex:1}}>
-        <LoginScreen1 />
+        <LoginScreen4 />
         </View>
     )
 }
